Adds the possiblity to use stateful sets stable network id
With the introduction of the StatefulSets, one may wish to use the
stable network ID of the pods, instead of their IPs. This change
adds one environmental variable 'KUBERNETES_MONGO_SERVICE_NAME' that
allows the user to enabled that feature, by explicitly defining a
(headless) mongo governing servie.

If this variable is present, the stable network id in the form will be
preferred:
`$(pod-name).$(service-name).$(namespace).svc.cluster.local`.

Extends the README.md to cover that new functionality.